Version 4.2.2
Fix a bug in UpdateAutoDewPower(int DewChannel)
Changed int DewPower = 0; to double DewPower=0;
If DewPower is not a double the DewPower1/2 calculation will either be 0 or 100.

+++ Arduino/Q-Astro-DewMonitor/ObservingConditions.ino
@@ -197,7 +197,7 @@ double GetSensorTemp(int sensor)
 void UpdateAutoDewPower(int DewChannel)
 {
   double Temp = GetSensorTemp(DewChannel);
-  int DewPower = 0;
+  double DewPower = 0;
 
   if (Temp != 99)
   {
